renamed TimePlacePickerFragement to PlacePickerFragmentActivity. Removed unused methods in Feed Activity. Set new mipmap
