+++ app/src/main/res/layout/fragment_add_location.xml
@@ -33,6 +33,7 @@
         android:layout_gravity="top|center"
         android:background="@color/colorPink"
         android:text="ADD LOCATION"
+        android:textSize="20dp"
         android:layout_weight="0.04" />
 
 </LinearLayout>
