+++ app/src/main/res/layout/fragment_place_picker.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical">
+
+    <fragment xmlns:app="http://schemas.android.com/apk/res-auto"
+        android:id="@+id/ad_fragment"
+        android:name="com.timecapsule.app.addmediafragment.PlacePickerFragmentActivity"
+        android:layout_width="fill_parent"
+        android:layout_height="50dp"
+        android:layout_alignParentBottom="true"
+        app:camera="camera" />
+
+</RelativeLayout>
