--- app/src/main/res/layout/fragment_time_place_picker.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <fragment
-        android:id="@+id/ad_fragment"
-        android:name="com.timecapsule.app.geofence.TimePlacePickerFragment"
-        android:layout_width="fill_parent"
-        android:layout_height="50dp"
-        android:layout_alignParentBottom="true"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        app:camera="camera"
-        />
-
-</RelativeLayout>
