+++ app/src/main/res/values/attrs.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <declare-styleable name="TimePlacePickerFragment">
+    <declare-styleable name="PlacePickerFragmentActivity">
         <attr name="camera" format="string"/>
         <attr name="video" format="string"/>
         <attr name="audio" format="string"/>
